Moves ChipSelect and ColorPicker
