+++ src/components/Field/FieldTypes.tsx
@@ -1,8 +1,8 @@
 import { FieldDefAddress } from "@root/components/Field/FormFieldAddress";
 import { FieldDefAdvancedSelection } from "@root/components/Field/FormFieldAdvancedSelection";
 import { FieldDefCheckbox } from "@root/components/Field/FormFieldCheckbox";
-import { FieldDefChip } from "@root/forms/FormFieldChipSingleSelect";
-import { FieldDefColor } from "@root/forms/FormFieldColorPicker/ColorPickerTypes";
+import { FieldDefChip } from "@root/components/Field/FormFieldChipSingleSelect";
+import { FieldDefColor } from "@root/components/Field/FormFieldColorPicker/ColorPickerTypes";
 import { FieldDefDate } from "@root/forms/FormFieldDate/DateField";
 import { FieldDefDropdown } from "@root/forms/FormFieldDropdownSingleSelection";
 import { FieldDefImageUpload } from "@root/forms/FormFieldImageUpload";
