+++ src/components/Field/FormFieldColorPicker/ColorPicker.stories.tsx
