+++ src/components/Form/Col/Col.tsx
@@ -5,8 +5,9 @@ import { formActions } from "../formActions";
 import FormFieldAddress from "@root/components/Field/FormFieldAddress";
 import FormFieldAdvancedSelection from "@root/components/Field/FormFieldAdvancedSelection";
 import FormFieldCheckbox from "@root/components/Field/FormFieldCheckbox";
+import FormFieldChipSingleSelect from "@root/components/Field/FormFieldChipSingleSelect";
+import FormFieldColorPicker from "@root/components/Field/FormFieldColorPicker";
 import FormFieldText from "@root/forms/FormFieldText";
-import FormFieldChipSingleSelect from "@root/forms/FormFieldChipSingleSelect";
 import FormFieldDropdownSingleSelection from "@root/forms/FormFieldDropdownSingleSelection";
 import FormFieldPhoneSelectionDropdown from "@root/forms/FormFieldPhoneSelectionDropdown";
 import FormFieldRadio from "@root/forms/FormFieldRadio";
@@ -14,7 +15,6 @@ import FormFieldRaw from "@root/forms/FormFieldRaw";
 import FormFieldToggleSwitch from "@root/forms/FormFieldToggleSwitch";
 import { FieldDef } from "@root/components/Field";
 import FormFieldImageVideoLinkDocumentBrowsing from "@root/forms/FormFieldImageVideoLinkDocumentBrowsing";
-import FormFieldColorPicker from "@root/forms/FormFieldColorPicker";
 import FormFieldDate from "@root/forms/FormFieldDate/DateField";
 import FormFieldTable from "@root/forms/FormFieldTable";
 import FormFieldTextEditor from "@root/forms/FormFieldTextEditor";
