+++ src/transforms/column_transforms.tsx
@@ -5,7 +5,7 @@ import { MosaicLabelValue, MosaicObject } from "../types";
 import { DataViewColumnTransform } from "../components/DataView";
 import Chip from "@root/components/Chip";
 import { ChipsWrapper, ColorValue } from "@root/components/Content/Content.styled";
-import ColorSelected from "@root/forms/FormFieldColorPicker/ColorSelected";
+import ColorSelected from "@root/components/Field/FormFieldColorPicker/ColorSelected";
 import Image from "@root/components/Image";
 import React from "react";
 import { DATE_FORMAT_SHORT } from "@root/constants";
